codex: address PR review feedback (#27783)
